pyproject.toml
[project]
name = "doc-genie"
version = "0.2.0"
description = "Extract black boxes from docs/code, map relationships, discover cross-doc insights"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "waynejing995"}]

dependencies = [
    # tree-sitter 核心 + 默认语言包 (C/JS/TS/Python)
    "tree-sitter>=0.24.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-c>=0.23.0",
    # 文档解析
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "markdown-it-py>=3.0.0",
    # 图分析
    "networkx>=3.0",
    # 工具
    "pyyaml>=6.0",
    "rich>=13.0",
    "reportlab>=4.4.10",
]

[project.optional-dependencies]
languages = [
    "tree-sitter-go>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-rust>=0.23.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[project.scripts]
genie = "scripts.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib"]
